+++ tests-lib/tget_constant_as_uint64.c
@@ -98,7 +98,7 @@ int main(void) {
   /* A constant close to an overflow */
   res = -17;
   a = SOLLYA_CONST(1);
-  for (i=1;(unsigned)(i)<=sizeof(int64_t)*8;i++) a = SOLLYA_MUL(a, SOLLYA_CONST(2));
+  for (i=1;(size_t)(i)<=sizeof(int64_t)*8;i++) a = SOLLYA_MUL(a, SOLLYA_CONST(2));
   a = SOLLYA_SUB(a, SOLLYA_CONST(1));
   if (!sollya_lib_get_constant_as_uint64(&res, a))
     sollya_lib_printf("%b is not a constant.\n\n", a);
@@ -111,7 +111,7 @@ int main(void) {
   /* A constant that overflows */
   res = -17;
   a = SOLLYA_CONST(1);
-  for (i=1;(unsigned)(i)<=sizeof(int64_t)*8;i++) a = SOLLYA_MUL(a, SOLLYA_CONST(2));
+  for (i=1;(size_t)(i)<=sizeof(int64_t)*8;i++) a = SOLLYA_MUL(a, SOLLYA_CONST(2));
   if (!sollya_lib_get_constant_as_uint64(&res, a))
     sollya_lib_printf("%b is not a constant.\n\n", a);
   else {
